+++ src/modules/system/migrate/scripts/index.js
@@ -19,3 +19,4 @@ export { default as InstalledItemMigrationFix } from "./012-installedItemMigrati
 export { default as RoleAbilitiesFix } from "./013-roleAbilitiesFix.js";
 export { default as SituationalFix } from "./014-situationalFix.js";
 export { default as dvCompendiumMigration } from "./015-dvCompendium.js";
+export { default as criticalInjuryCompendiumMigration } from "./017-critTables.js";