+++ src/modules/system/migrate/scripts/013-roleAbilitiesFix.js
@@ -1,3 +1,5 @@
+/* eslint-disable foundry-cpr/logger-after-function-definition */
+
 import CPRMigration from "../cpr-migration.js";
 import LOGGER from "../../../utils/cpr-logger.js";