+++ src/modules/item/sheet/cpr-item-sheet.js
@@ -86,10 +86,7 @@ export default class CPRItemSheet extends ItemSheet {
 
     // Enrich the description so that links to foundry documents in item descriptions have proper functionality.
     foundryData.enrichedHTMLDescription = await TextEditor.enrichHTML(
-      foundryData.item.system.description.value,
-      {
-        async: true,
-      }
+      foundryData.item.system.description.value
     );
     return foundryData;
   }