+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 
 ### Bug Fixes
 
-- #808 - Installed items in a mook were mapped to the wrong mook _id, this has been fixed.
+- #808 - Installed items in a mook were mapped to the wrong mook \_id, this has been fixed.
 - #812 - Some roles were missing the bonuses data point as an empty array.
+- Fix being unable to delete certain effects from compendium items with Active Effects.
 
 ## Version 0.87.2