+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - #812 - Some roles were missing the bonuses data point as an empty array.
 - Fix being unable to delete certain effects from compendium items with Active Effects.
 - Fix CSS issues with modules Cautious Gamemaster's Pack and Item Piles
+- Fix #824 - Now when you drag a cyberware item to a mook sheet, if it can't be auto-installed/equipped, it will delete the newly created item so it does not remain hidden in the inventory.
 
 ## Version 0.87.2