+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 - #812 - Some roles were missing the bonuses data point as an empty array.
 - Fix being unable to delete certain effects from compendium items with Active Effects.
 - Fix CSS issues with modules Cautious Gamemaster's Pack and Item Piles
-- Fix #824 - Now when you drag a cyberware item to a mook sheet, if it can't be auto-installed/equipped, it will delete the newly created item so it does not remain hidden in the inventory.
+- #824 - Now when you drag a cyberware item to a mook sheet, if it can't be auto-installed/equipped, it will delete the newly created item so it does not remain hidden in the inventory.
 - Fix Crushed Fingers/Lost Eye Critical Injury Effect Name
+- #827 - Fix items not being created when purchased from a vendor/container actor.
 
 ## Version 0.87.2