+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - #827 - Fix items not being created when purchased from a vendor/container actor.
 - #829 - Fix failure when migrating tokens with broken references to their parent actors.
 - Fix issue where unlinked actors were not gaining Active Effects.
+- Fix compendia name references for critical injuries which was preventing rolling critical injuries
 
 ## Version 0.87.2