+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - #846 - Installed items losing their installation data after migration
 - Various minor CSS fixes
 - #854 - Fix missing Babele translations for Critical Injuries preventing rolling Critical Injuries from character sheets
+- Fix rolling custom skills.
 
 ## Version 0.87.2