+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 # Changelog
 
+## Version 0.88.0
+
+### Changes
+
+- Update System files for v11
+- Update `label` > `name` for Active effects
+- Hide system only Compendiums from Compendium tab
+- Organize Compendiums into COmpendium Folders
+
 ## Version 0.87.3
 
 ### Bug Fixes